CLOUDP-121809: ask to regenerate device code on login flow
