+++ internal/cli/auth/login.go
@@ -16,11 +16,11 @@ package auth
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"os"
 	"time"
 
+	"github.com/AlecAivazis/survey/v2"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli"
 	"github.com/mongodb/mongodb-atlas-cli/internal/cli/require"
 	"github.com/mongodb/mongodb-atlas-cli/internal/config"
@@ -52,18 +52,44 @@ const (
 	LogoutToLoginAccountMsg      = `run "atlas auth logout" first if you want to login with another Atlas account on the same Atlas CLI profile`
 )
 
-var errTimedOut = errors.New("authentication timed out")
-
 type LoginOpts struct {
 	cli.DefaultSetterOpts
-	AccessToken    string
-	RefreshToken   string
-	IsGov          bool
-	isCloudManager bool
-	NoBrowser      bool
-	SkipConfig     bool
-	config         LoginConfig
-	flow           Authenticator
+	AccessToken          string
+	RefreshToken         string
+	IsGov                bool
+	isCloudManager       bool
+	NoBrowser            bool
+	SkipConfig           bool
+	config               LoginConfig
+	flow                 Authenticator
+	regenerateCodePrompt userSurvey
+}
+
+func NewLoginOpts() *LoginOpts {
+	return &LoginOpts{
+		regenerateCodePrompt: &confirmPrompt{
+			message:         "Your one-time verification code is expired. Would you like to generate a new one?",
+			defaultResponse: true,
+		},
+	}
+}
+
+type userSurvey interface {
+	confirm() (response bool, err error)
+}
+
+type confirmPrompt struct {
+	message         string
+	defaultResponse bool
+}
+
+func (c *confirmPrompt) confirm() (response bool, err error) {
+	p := &survey.Confirm{
+		Message: c.message,
+		Default: c.defaultResponse,
+	}
+	err = survey.AskOne(p, &response)
+	return response, err
 }
 
 type LoginFlow interface {
@@ -174,30 +200,46 @@ Your code will expire after %.0f minutes.
 }
 
 func (opts *LoginOpts) oauthFlow(ctx context.Context) error {
-	code, _, err := opts.flow.RequestCode(ctx)
-	if err != nil {
-		return err
-	}
-
-	opts.printAuthInstructions(code)
-
-	if !opts.NoBrowser {
-		if errBrowser := browser.OpenURL(code.VerificationURI); errBrowser != nil {
-			_, _ = fmt.Fprintf(os.Stderr, "There was an issue opening your browser\n")
-		}
-	}
-
-	var accessToken *auth.Token
-	if accessToken, _, err = opts.flow.PollToken(ctx, code); err != nil {
-		if auth.IsTimeoutErr(err) {
-			return errTimedOut
+	browserOpened := false
+	for {
+		code, _, err := opts.flow.RequestCode(ctx)
+		if err != nil {
+			return err
 		}
 
-		return err
+		opts.printAuthInstructions(code)
+
+		if !opts.NoBrowser && !browserOpened {
+			if errBrowser := browser.OpenURL(code.VerificationURI); errBrowser != nil {
+				_, _ = fmt.Fprintf(os.Stderr, "There was an issue opening your browser\n")
+			} else {
+				browserOpened = true
+			}
+		}
+
+		accessToken, _, err := opts.flow.PollToken(ctx, code)
+		if retry, errRetry := opts.shouldRetryRegister(err); errRetry != nil {
+			return errRetry
+		} else if retry {
+			continue
+		}
+
+		if err != nil {
+			return err
+		}
+
+		opts.AccessToken = accessToken.AccessToken
+		opts.RefreshToken = accessToken.RefreshToken
+		return nil
 	}
-	opts.AccessToken = accessToken.AccessToken
-	opts.RefreshToken = accessToken.RefreshToken
-	return nil
+}
+
+func (opts *LoginOpts) shouldRetryRegister(err error) (retry bool, errSurvey error) {
+	if err == nil || !auth.IsTimeoutErr(err) {
+		return false, nil
+	}
+
+	return opts.regenerateCodePrompt.confirm()
 }
 
 func hasUserProgrammaticKeys() bool {
@@ -239,7 +281,7 @@ func Tool() string {
 }
 
 func LoginBuilder() *cobra.Command {
-	opts := &LoginOpts{}
+	opts := NewLoginOpts()
 
 	cmd := &cobra.Command{
 		Use:   "login",
